requirements.txt
opencv-python>=4.8.0
zxing-cpp>=2.2.0
pywebview>=5.0
numpy>=1.24.0
pillow>=10.0
pillow-avif-plugin>=1.4.0
